Splitting the Iranian Agent Department
