+++ TheInvestigationGame_15_06/Main/GameManager.cs
@@ -8,27 +8,34 @@ namespace TheInvestigationGame_15_06
     internal class GameManager
     {
         private IranianAgent currentAgent;
+
         public void StartGame()
         {
             Console.WriteLine("Welcome to The Investigation Game!");
             Console.WriteLine("Starting investigation...\n");
-            // פה יכנסו סוכנים
+
+            // Create agent
             currentAgent = new FootSoldier();
-            while (!currentAgent.IsRevealed())
+
+            // Loop until all weaknesses are revealed
+            while (!currentAgent.sensorManager.IsRevealed())
             {
                 ShowAvailableSensors();
                 HandleSensorChoice();
             }
+
             Console.WriteLine("The agent has been fully exposed!");
         }
+
         private void ShowAvailableSensors()
         {
             Console.WriteLine("Available Sensors:");
-            for (int i = 0; i < IranianAgent.allSensors.Count; i++)
+            for (int i = 0; i < SensorManager.allSensors.Count; i++)
             {
-                Console.WriteLine($"{i + 1}. {IranianAgent.allSensors[i].Name}");
+                Console.WriteLine($"{i + 1}. {SensorManager.allSensors[i].Name}");
             }
         }
+
         private void HandleSensorChoice()
         {
             Console.Write("Choose a sensor by number: ");
@@ -37,8 +44,8 @@ namespace TheInvestigationGame_15_06
             int choice;
             if (ValidChoice(input, out choice))
             {
-                Sensor selectedSensor = IranianAgent.allSensors[choice - 1];
-                string result = currentAgent.ActivateSensor(selectedSensor);
+                Sensor selectedSensor = SensorManager.allSensors[choice - 1];
+                string result = currentAgent.sensorManager.ActivateSensor(selectedSensor, currentAgent);
                 Console.WriteLine(result);
             }
             else
@@ -46,6 +53,7 @@ namespace TheInvestigationGame_15_06
                 Console.WriteLine("Unvalid choice. Please try again.");
             }
         }
+
         private bool ValidChoice(string input, out int choice)
         {
             bool isNumber = int.TryParse(input, out choice);
@@ -53,12 +61,13 @@ namespace TheInvestigationGame_15_06
             {
                 return false;
             }
-            if (choice < 1 || choice > IranianAgent.allSensors.Count)
+            if (choice < 1 || choice > SensorManager.allSensors.Count)
             {
                 return false;
             }
             return true;
         }
+
     }
 }
