+++ TheInvestigationGame_15_06/Sensors/ThermalSensor.cs
@@ -10,27 +10,24 @@ namespace TheInvestigationGame_15_06.Sensors
     internal class ThermalSensor : Sensor
     {
         public override string Name => "Thermal Sensor";
+
         public override string Activate(IranianAgent agent)
         {
             return GetAnotherWeakness(agent);
-
         }
+
         private static string GetAnotherWeakness(IranianAgent agent)
         {
-            foreach (Sensor secret in agent.secretSensors)
+            foreach (Sensor secret in agent.sensorManager.secretSensors)
             {
-                if (!agent.IsSensorRevealed(secret))
-                {
                     string result = "Thermal Sensor activated!";
-                    if (!agent.IsRevealed())
+                    if (!agent.sensorManager.IsRevealed())
                     {
                         result += $" Get another weakness: {secret.Name}";
                     }
                     return result;
-                }
             }
-            return "Thermal Sensor activate";
+            return "Thermal Sensor activated!";
         }
-
     }
 }
